Pool additions
Acquiring a connection now creates a new one if necessary, if there is
room left to grow the pool. Non-connected connections are pruned from
the pool when acquiring a new connection as well.

When attempting to acquire a connection if none are available and there
is no more room left to grow the pool a wait is enforced until a
connection is freed up.

When releasing a connection back into the pool non-connected connections
are pruned.

Using the pool as an async context (`async with pool as conn`) acquires
a connection and puts it into pipeline mode.

A single-shot `execute` acquires a connection for the execution.