Fixed an issue involving connection pipeline future resolution.
Futures are wrapped while pipelining, and in the event a connection had
an ordinary future and then went into pipeline mode and wrapped
all subsequent futures when exiting the pipeline it attempted to call a
method on all futures assuming they were all wrapped throwing an
`AttributeError`.

There's a check now on all resolvers and only if they originated from
within a pipeline is the wrapper method called.